fix qst and test it add event but add stimulus
