+++ poulet_py/hardware/stimulator/qst/__init__.py
@@ -1,0 +1,3 @@
+from lazy_loader import attach_stub
+
+__getattr__, __dir__, __all__ = attach_stub(__name__, __file__)
