+++ poulet_py/io/__init__.pyi
@@ -1,15 +1,17 @@
 # ruff: noqa TID252
 from .sinks import BaseSink, HDFSink
-from .sources import BaseSource, NIAnalogInputSource, CounterSource, TCSSource
-from .data_packets import BaseDataPacket
+from .sources import BaseSource, NIAnalogInputSource, CounterSource, TCSSource, AcquisitionType
+from .eventing import EventBus, Event
 
 __all__ = [
     "BaseSink",
     "HDFSink",
     "BaseSource",
     "NIAnalogInputSource",
-    "BaseDataPacket",
+    "EventBus",
+    "Event",
     "HDFSink",
     "CounterSource",
     "TCSSource",
+    "AcquisitionType",
 ]
