--- poulet_py/io/data_packets/__init__.pyi
@@ -1,4 +1,0 @@
-# ruff: noqa TID252
-from .common import BaseDataPacket
-
-__all__ = ["BaseDataPacket"]
