+++ poulet_py/io/eventing/__init__.pyi
@@ -1,0 +1,5 @@
+# ruff: noqa TID252
+from .events import Event
+from .event_bus import EventBus
+
+__all__ = ["EventBus", "Event"]
