+++ poulet_py/io/sources/__init__.pyi
@@ -1,7 +1,7 @@
 # ruff: noqa TID252
-from .common import BaseSource
+from .common import BaseSource, AcquisitionType
 from .nidaq import NIAnalogInputSource
 from .counter import CounterSource
 from .tcs import TCSSource
 
-__all__ = ["BaseSource", "NIAnalogInputSource", "CounterSource", "TCSSource"]
+__all__ = ["BaseSource", "NIAnalogInputSource", "CounterSource", "TCSSource", "AcquisitionType"]
