+++ poulet_py/tools/__init__.pyi
@@ -1,12 +1,12 @@
 # ruff: noqa TID252
-from .generators import generate_stimulus_sequence
+from .generators import repeat
 from .organizational import check_or_create, define_folder_name, go_to, sanitize_path
 from .serializers import json_serializer, save_metadata_exp
 
 __all__ = [
     "check_or_create",
     "define_folder_name",
-    "generate_stimulus_sequence",
+    "repeat",
     "go_to",
     "json_serializer",
     "sanitize_path",
